backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlmodel>=0.0.22
python-multipart>=0.0.18
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.28.0
aiofiles>=24.1.0
jinja2>=3.1.0
pyyaml>=6.0
